set content type to application/json

+++ src/main/java/restApp/Logic/Fibonacci.java
@@ -9,10 +9,10 @@ public class Fibonacci {
     returns the fibonacci number at position n
     @param n represents the position of the fibonacci number
      */
-    public static Integer getFibonacciNumberAt(Integer n) throws ArithmeticException
+    public static long getFibonacciNumberAt(long n) throws ArithmeticException
     {
-        Integer a = 0;
-        Integer b = 1;
+        long a = 0;
+        long b = 1;
 
         if(n < 0){
             return generateFibonacciNegative(-a,-b,n);
@@ -23,7 +23,7 @@ public class Fibonacci {
     /*
     helper function for getFibonacciNumberAt
      */
-    private static Integer generateFibonacciPositive(int a, int b, int n) throws ArithmeticException{
+    private static long generateFibonacciPositive(long a, long b, long n) throws ArithmeticException{
         if(n == 0){
             return a;
         }
@@ -31,7 +31,7 @@ public class Fibonacci {
         return generateFibonacciPositive(b, Math.addExact(a,b), n-1);
     }
 
-    private static Integer generateFibonacciNegative(int a, int b, int n) throws ArithmeticException{
+    private static long generateFibonacciNegative(long a, long b, long n) throws ArithmeticException{
 
         if(n == 0){
             return a;
